DEV-3872: Abstract Contact Info and add tests
