+++ spa/src/components/portal/ContributionsList/ContactInfoPopover/__mocks__/ContactInfoPopover.tsx
@@ -1,0 +1,7 @@
+import { ContactInfoPopoverProps } from '../ContactInfoPopover';
+
+const ContactInfoPopover = ({ page }: ContactInfoPopoverProps) => (
+  <div data-testid="mock-contact-info-popover" data-page={JSON.stringify(page)} />
+);
+
+export default ContactInfoPopover;
